requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.7.14
click==8.2.2
deprecation==2.1.0
fastapi==0.104.1
gotrue==1.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
packaging==25.0
postgrest==0.13.2
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
realtime==1.0.6
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.2
supafunc==0.3.3
typing_extensions==4.14.1
uvicorn==0.24.0
websockets==12.0
